Main Api Completed

+++ src/MedSino.Service/Services/Users/UserService.cs
@@ -1,15 +1,16 @@
 ﻿using MedSino.DataAccess.Interfaces;
 using MedSino.DataAccess.Interfaces.Users;
 using MedSino.DataAccess.Utils;
-using MedSino.Domain.Entities.Doctors;
+using MedSino.Domain.Entities.Categories;
 using MedSino.Domain.Entities.Users;
 using MedSino.Domain.Exceptions.Categories;
+using MedSino.Domain.Exceptions.Files;
 using MedSino.Domain.Exceptions.Users;
 using MedSino.Service.Common.Helpers;
 using MedSino.Service.Dtos.Users;
+using MedSino.Service.Interfaces.Auth;
 using MedSino.Service.Interfaces.Common;
 using MedSino.Service.Interfaces.Users;
-using MedSino.Service.Services.Common;
 
 namespace MedSino.Service.Services.Users;
 
@@ -18,20 +19,28 @@ public class UserService : IUserService
     private readonly IUserRepository _userRepository;
     private readonly IFileService _fileService;
     private readonly IPaginator _paginator;
+    private readonly IIdentityService _identity;
 
     public UserService(IUserRepository userRepository,
         IFileService fileService,
-        IPaginator paginator)
+        IPaginator paginator,
+        IIdentityService identityService)
     {
         this._userRepository = userRepository;
         this._fileService = fileService;
         this._paginator = paginator;
+        this._identity = identityService;
     }
 
     public async Task<bool> DeleteAsync(long userId)
     {
-        var result = await _userRepository.DeleteAsync(userId);
-        return result > 0;
+        //if(userId != _identity.UserId && _identity.IdentityRole != Domain.Enums.IdentityRole.Admin) throw new UserNotFoundException();
+        var user = await _userRepository.GetByIdAsync(userId);
+        if (user is null) throw new UserNotFoundException();
+
+        var result = await _fileService.DeleteImageAsync(user.ImagePath);
+        var dbResult = await _userRepository.DeleteAsync(userId);
+        return dbResult > 0;
     }
 
     public async Task<IList<User>> GetAllAsync(PaginationParams @params)
@@ -51,8 +60,9 @@ public class UserService : IUserService
 
     public async Task<bool> UpdateAsync(long userId, UserUpdateDto dto)
     {
+        //if (userId != _identity.UserId && _identity.IdentityRole != Domain.Enums.IdentityRole.Admin) throw new UserNotFoundException();
         var user = await _userRepository.GetByIdAsync(userId);
-        if (user is null) throw new CategoryNotFoundException();
+        if (user is null) throw new UserNotFoundException();
 
         user.FirstName = dto.FirstName;
         user.LastName = dto.LastName;
